requirements.txt
pixelblaze-client
python-weather